Implement SOCKS5 proxy support with CONNECT and UDP ASSOCIATE methods
- Added `pb_socks5.c` for handling SOCKS5 protocol operations including:
  - `socks5_read_connect_reply`: Reads and validates SOCKS5 CONNECT replies.
  - `socks5_connect_domain`: Establishes a SOCKS5 connection using a domain name.
  - `socks5_connect`: Handles SOCKS5 connections using IPv4 addresses.
  - `socks5_connect_v6`: Handles SOCKS5 connections using IPv6 addresses.
  - `socks5_udp_associate_with_config`: Manages UDP association with the SOCKS5 proxy.
  - `establish_udp_associate_for_config`: Establishes UDP association for the given proxy configuration.

- Added `pb_util.c` for utility functions including:
  - Logging functionality with `log_message`.
  - IP address formatting and parsing functions.
  - Socket configuration functions for TCP and UDP.
  - Functions for connecting with a timeout and sending/receiving data reliably.
  - DNS resolution for hostnames to IPv4 addresses.
  - Base64 encoding utility.

These changes enhance the proxy capabilities and improve the handling of network communications.
